Separated out of App.tsx the table data structures into individual files to declutter App.tsx
