+++ src/App.tsx
@@ -12,6 +12,11 @@ import { Table } from 'antd';
 import { AreaChart, Area, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts';
 import SearchBar from './SearchBar'
 import type { ColumnsType } from 'antd/es/table';
+import myStashColumns from './myStashColumns';
+import snapshotColumns from './snapshotColumns';
+import allStocksColumns from './allStocksColumns';
+import hourlyColumns from './hourlyColumns';
+
 
 
 const App = () => {
@@ -56,179 +61,6 @@ const App = () => {
     }
 
 
-    interface TableDataType {
-        key: React.Key;
-        name: string;
-        symbol: string;
-        price: number;
-        eps: string;
-        dayHigh: number;
-        dayLow: number;
-        change: number;
-        date: string;
-        open: number;
-        low: number;
-        high: number;
-        close: number;
-        volume: number;
-      }
-
-
-    const myStashColumns: ColumnsType<TableDataType> = [
-    {
-        title: 'Symbol',
-        dataIndex: 'symbol',
-        key: 'symbol'
-    },
-    {
-        title: 'Name',
-        dataIndex: 'name',
-        key: 'name',
-        render: text => <a>{text}</a>
-    },
-    {
-        title: 'Price',
-        dataIndex: 'price',
-        key: 'price'
-    },
-    {
-        title: 'EPS',
-        dataIndex: 'eps',
-        key: 'eps'
-    },
-    {
-        title: 'DayHigh',
-        dataIndex: 'dayHigh',
-        key: 'dayHigh'
-    },
-    {
-        title: 'DayLow',
-        dataIndex: 'dayLow',
-        key: 'dayLow'
-    },
-    {
-        title: 'Change',
-        dataIndex: 'change',
-        key: 'change'
-    },
-    {
-        title: 'Unstash',
-        dataIndex: 'unstash',
-        key: 'unstash',
-        render: Unstash => <button onClick={() => unstash}>Remove</button>
-    }
-    ];
-      
-
-    const snapshotColumns: ColumnsType<TableDataType> = [
-    {
-        title: 'Symbol',
-        dataIndex: 'symbol',
-        key: 'symbol'
-    },
-    {
-        title: 'Name',
-        dataIndex: 'name',
-        key: 'name',
-        render: text => <a>{text}</a>
-    },
-    {
-        title: 'Price',
-        dataIndex: 'price',
-        key: 'price'
-    },
-    {
-        title: 'EPS',
-        dataIndex: 'eps',
-        key: 'eps'
-    },
-    {
-        title: 'DayHigh',
-        dataIndex: 'dayHigh',
-        key: 'dayHigh'
-    },
-    {
-        title: 'DayLow',
-        dataIndex: 'dayLow',
-        key: 'dayLow'
-    },
-    {
-        title: 'Change',
-        dataIndex: 'change',
-        key: 'change'
-    }
-    ];
-
-
-    const allStocksColumns: ColumnsType<TableDataType> = [
-    {
-        title: 'Symbol',
-        dataIndex: 'symbol',
-        key: 'symbol'
-    },
-    {
-        title: 'Name',
-        dataIndex: 'name',
-        key: 'name',
-        render: text => <a>{text}</a>
-    },
-    {
-        title: 'Price',
-        dataIndex: 'price',
-        key: 'price'
-    },
-    {
-        title: 'Exchange',
-        dataIndex: 'exchange',
-        key: 'exchange'
-    },
-    {
-        title: 'ExchangeShortName',
-        dataIndex: 'exchangeShortName',
-        key: 'exchangeShortName'
-    },
-    {
-        title: 'Type',
-        dataIndex: 'type',
-        key: 'type'
-    }
-    ];
-
-
-    const hourlyColumns: ColumnsType<TableDataType> = [
-        {
-            title: 'Close',
-            dataIndex: 'close',
-            key: 'close'
-        },
-        {
-            title: 'Date',
-            dataIndex: 'date',
-            key: 'date'
-        },
-        {
-            title: 'High',
-            dataIndex: 'high',
-            key: 'high'
-        },
-        {
-            title: 'Low',
-            dataIndex: 'low',
-            key: 'low'
-        },
-        {
-            title: 'Open',
-            dataIndex: 'open',
-            key: 'open'
-        },
-        {
-            title: 'Volume',
-            dataIndex: 'volume',
-            key: 'volume'
-        }
-        ];
-
-    
       
     const snapshotTable = () => <Table className="flex-container" columns={snapshotColumns} dataSource={stockResults} />;
     const myStashTable = () => <Table className="flex-container" columns={myStashColumns} dataSource={stockStash} />;
